+++ packages/contracts-communication/contracts/interfaces/IInterchainApp.sol
@@ -13,7 +13,7 @@ interface IInterchainApp {
     function appReceive(
         uint64 srcChainId,
         bytes32 sender,
-        uint256 dbNonce,
+        uint64 dbNonce,
         uint64 entryIndex,
         bytes calldata message
     )
